# engine/docs2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flapjack API",
    "description": "Drop-in replacement for Algolia search. Provides full-text search with filters, facets, geo-search, and more.",
    "contact": {
      "name": "Flapjack",
      "url": "https://github.com/griddlehq/flapjack"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7700",
      "description": "Local development"
    },
    {
      "url": "https://fj-us-west-1.flapjack.foo",
      "description": "Production US West"
    }
  ],
  "paths": {
...
    "/1/indexes/{indexName}/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get index settings",
        "operationId": "get_settings",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Index settings",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Index not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "operationId": "set_settings_put_doc",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Settings to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid settings"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Update the settings for the specified index.",
        "operationId": "set_settings",
        "parameters": [
          {
            "name": "indexName",
            "in": "path",
            "description": "Index name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Settings to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid settings"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IgnorePluralsValue": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "Disabled"
            ]
          },
          {
            "type": "string",
            "enum": [
              "All"
            ]
          },
          {
            "type": "object",
            "required": [
              "Languages"
            ],
            "properties": {
              "Languages": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "IndexMode": {
        "type": "string",
        "enum": [
          "keywordSearch",
          "neuralSearch"
        ]
      },
...
      "RemoveStopWordsValue": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "Disabled"
            ]
          },
          {
            "type": "string",
            "enum": [
              "All"
            ]
          },
          {
            "type": "object",
            "required": [
              "Languages"
            ],
            "properties": {
              "Languages": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
      "SemanticSearchSettings": {
        "type": "object",
        "properties": {
          "eventSources": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SetSettingsRequest": {
        "allOf": [
          {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          {
            "type": "object",
            "properties": {
              "advancedSyntaxFeatures": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "allowCompressionOfIntegerArray": {
                "type": [
                  "boolean",
                  "null"
                ]
              },
              "attributeCriteriaComputedByMinProximity": {
                "type": [
                  "boolean",
                  "null"
                ]
              },
              "attributeForDistinct": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "attributesForFaceting": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "attributesToRetrieve": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "customRanking": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "disableExactOnAttributes": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "disableTypoToleranceOnAttributes": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "disableTypoToleranceOnWords": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "distinct": {},
              "embedders": {
                "type": [
                  "object",
                  "null"
                ],
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              },
              "enablePersonalization": {
                "type": [
                  "boolean",
                  "null"
                ]
              },
              "enableReRanking": {
                "type": [
                  "boolean",
                  "null"
                ]
              },
              "enableRules": {
                "type": [
                  "boolean",
                  "null"
                ]
              },
              "ignorePlurals": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/IgnorePluralsValue"
                  }
                ]
              },
              "minProximity": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int32",
                "minimum": 0
              },
              "mode": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/IndexMode"
                  }
                ]
              },
              "numericAttributesForFiltering": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "paginationLimitedTo": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int32",
                "minimum": 0
              },
              "queryLanguages": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "queryType": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "ranking": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "relevancyStrictness": {
                "type": [
                  "integer",
                  "null"
                ],
                "format": "int32",
                "minimum": 0
              },
              "removeStopWords": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RemoveStopWordsValue"
                  }
                ]
              },
              "renderingContent": {},
              "replaceSynonymsInHighlight": {
                "type": [
                  "boolean",
                  "null"
                ]
              },
              "replicas": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "restrictHighlightAndSnippetArrays": {
                "type": [
                  "boolean",
                  "null"
                ]
              },
              "searchableAttributes": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "semanticSearch": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/SemanticSearchSettings"
                  }
                ]
              },
              "snippetEllipsisText": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "sortFacetValuesBy": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "unretrievableAttributes": {
                "type": [
                  "array",
                  "null"
                ],
                "items": {
                  "type": "string"
                }
              },
              "userData": {}
            }
          }
        ],
        "description": "Deserializable payload for the `POST /1/indexes/{indexName}/settings` endpoint.\n\nAll fields are optional; only supplied fields are merged into the existing `IndexSettings`.\nUnrecognized keys are captured in `other` via `#[serde(flatten)]` and reported as unsupported."
      },
      "SetSettingsResponse": {
        "type": "object",
        "required": [
          "updatedAt",
          "taskID"
        ],
        "properties": {
          "taskID": {
            "type": "integer",
            "format": "int64"
          },
          "unsupportedParams": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-algolia-api-key"
      }
    }
  },
...
}